refactor: extract message filter, remove dead code
- Extract toolOutputPatterns and isToolOutputNoise to
  packages/web/src/lib/message-filter.ts (was inline in App.tsx)
- Remove unused connectionsRef (superseded by getSessionIdRef)
- Remove no-op empty block in hello_ack handler
- Remove unused AppState interface from types/index.ts
- Remove stale relay comment
